Move snapshots to single file where the expected is seperate
